Added check to see if rvm-prompt exists.
I did this so that if it isn't found it simply uses a right prompt without the
rvm-prompt portion. However, if the rvm-prompt command is found then it
includes it in the right prompt. This prevents annoying failures to show up as
a result of the prompt of the user doesn't have RVM installed in their home
directory.