Added a "Why use this .zsh setup" section.
I did this so that when people look at this repo they can see what it is going
to provide them before they actually go to the effort of installing it. It
also doubles as a reference for people to understand what the different prompt
values represent and their various states. I also added a Credit section to
give credit to the various places that I snagged parts of this .zsh setup
from.