Split zshrc into zshrc & zshenv.
I did this because I learned that in non-interactive zsh shells it only loads
/etc/zshenv and then ~/.zshenv. It does NOT load /etc/zshrc or ~/.zshrc. This
is a problem for example in Vim when you run a command with :! and you want
Vim to have your aliases, RVM loadded properly, the proper path configured,
etc.

So, I basically moved everything that I wanted setup in all my shells
interactive, non-interactive, and login shells all into the zshenv file so
that all those things are loaded in all those shell types. I then left all the
stuff I only want to exist in interactive shells in the zshrc file as that
only gets loaded in the interactive shells.