Removed xterm 256 color override.
I did this because it was overriding the default terminal setting provided by
tmux which was causing weird color issues inside Vim when run in Tmux. Plus,
there is no point in overriding it because it is acutally being set properly
in the normal terminal and in the tmux session.