Make the rbenv path explicit.
I did this because the Mac OS X 10.10 Yosemite developer preview broke this
seemingly because the sub-shell doesn't have /usr/local/bin in its path any
longer. I am guessing it is because they either changed the default
environment or the shell that is used all together changed in the default
scenario.