Add rbenv to zshenv
Why you made the change:

I did this to get rbenv loaded in all shells.

How the change addresses the need:

This enables this because the $ZDOTDIR/.zshenv file is loaded for all
shells. You will note I also added logic to check if it has already been
loaded in the shell and to ignore it if it has.