Change rbenv-prompt to graceful fail
Why you made the change:

I did this so that if you have your shell configured in such a way that
rbenv-prompt is being used and you don't have rbenv installed/accessible
it will simply emit "rbenv is missing" in place of the ruby version in
your prompt. This prevents it from interfering with running of certain
programs buy causing shell init to be loud.