Disable PATH alterations in .zshenv
I did this because since macOS added the use path_helper in the
/etc/zprofile it prefixes the PATH with whatever paths are in /etc/paths
or /etc/paths.d. By moving the PATH modifications to the .zshrc it makes
the path correct as the modifications are happening after the
path_helper modifications. The downside to this is that non-interactive
shells won't have the PATH modifications. I am opting for general usage
case over the specialized scripting case.